requirements.txt
# PyTorch和相关库
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 深度学习工具
timm>=0.9.0
einops>=0.6.1
thop>=0.1.1

# 数据处理
numpy>=1.24.0
Pillow>=9.5.0

# 数据增强
albumentations>=1.3.1

# 评估指标
scikit-learn>=1.2.2
scipy>=1.10.1

# 可视化
matplotlib>=3.7.1
seaborn>=0.12.2

# 进度条和日志
tqdm>=4.65.0

# 配置管理
pyyaml>=6.0

# 其他工具
pandas>=2.0.0
gdal>=3.6.0  # 遥感数据读取(可选)